Rename package to nobleo_socketcan_bridge
