+++ src/socketcan_bridge_ee.cpp
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include "nobleo_socketcan_bridge/socketcan_bridge_node.hpp"
 #include "rclcpp/experimental/executors/events_executor/events_executor.hpp"
-#include "ros2_socketcan_bridge/socketcan_bridge_node.hpp"
 
-using ros2_socketcan_bridge::SocketCanBridgeNode;
+using nobleo_socketcan_bridge::SocketCanBridgeNode;
 
 int main(int argc, char * argv[])
 {
